+++ src/components/common/workspace/AddWorkspace.tsx
@@ -3,7 +3,7 @@ import { Workspace } from '@@types/index';
 import useAdminUser from '@hooks/admin/useAdminUser';
 import { useRef, useState } from 'react';
 import CreateWorkspaceModal from './modal/CreateWorkspaceModal';
-import AddWorkspaceModalContent from '../content/AddworkspaceModalContent';
+import AddWorkspaceModalContent from '../../admin/workspace/AddworkspaceModalContent';
 import PlusIconSvg from '@resources/svg/PlusIconSvg';
 
 const AddWorkspaceContainer = styled.form`
